Split file fetch.rs
